move proxy to new structure; shuffle util functions to better location
