+++ migration/versions.go
@@ -1,30 +1,8 @@
 package migration
 
 import (
-	"errors"
 	"github.com/coredns/corefile-migration/migration/corefile"
 )
-
-type plugin struct {
-	status         string
-	replacedBy     string
-	additional     string
-	namedOptions   map[string]option
-	patternOptions map[string]option
-	action         pluginActionFn // action affecting this plugin only
-	add            serverActionFn // action to add a new plugin to the server block
-	downAction     pluginActionFn // downgrade action affecting this plugin only
-}
-
-type option struct {
-	name       string
-	status     string
-	replacedBy string
-	additional string
-	action     optionActionFn // action affecting this option only
-	add        pluginActionFn // action to add the option to the plugin
-	downAction optionActionFn // downgrade action affecting this option only
-}
 
 type release struct {
 	k8sReleases    []string
@@ -48,54 +26,7 @@ type release struct {
 	defaultConf string
 }
 
-type corefileAction func(*corefile.Corefile) (*corefile.Corefile, error)
-type serverActionFn func(*corefile.Server) (*corefile.Server, error)
-type pluginActionFn func(*corefile.Plugin) (*corefile.Plugin, error)
-type optionActionFn func(*corefile.Option) (*corefile.Option, error)
-
-func removePlugin(*corefile.Plugin) (*corefile.Plugin, error) { return nil, nil }
-func removeOption(*corefile.Option) (*corefile.Option, error) { return nil, nil }
-
-func renamePlugin(p *corefile.Plugin, to string) (*corefile.Plugin, error) {
-	p.Name = to
-	return p, nil
-}
-
-func addToServerBlockWithPlugins(sb *corefile.Server, newPlugin *corefile.Plugin, with []string) (*corefile.Server, error) {
-	if len(with) == 0 {
-		// add to all blocks
-		sb.Plugins = append(sb.Plugins, newPlugin)
-		return sb, nil
-	}
-	for _, p := range sb.Plugins {
-		for _, w := range with {
-			if w == p.Name {
-				// add to this block
-				sb.Plugins = append(sb.Plugins, newPlugin)
-				return sb, nil
-			}
-		}
-	}
-	return sb, nil
-}
-
-func addToKubernetesServerBlocks(sb *corefile.Server, newPlugin *corefile.Plugin) (*corefile.Server, error) {
-	return addToServerBlockWithPlugins(sb, newPlugin, []string{"kubernetes"})
-}
-
-func addToForwardingServerBlocks(sb *corefile.Server, newPlugin *corefile.Plugin) (*corefile.Server, error) {
-	return addToServerBlockWithPlugins(sb, newPlugin, []string{"forward", "proxy"})
-}
-
-func addToAllServerBlocks(sb *corefile.Server, newPlugin *corefile.Plugin) (*corefile.Server, error) {
-	return addToServerBlockWithPlugins(sb, newPlugin, []string{})
-}
-
-func addOptionToPlugin(pl *corefile.Plugin, newOption *corefile.Option) (*corefile.Plugin, error) {
-	pl.Options = append(pl.Options, newOption)
-	return pl, nil
-}
-
+// Versions holds a map of plugin/option migrations per CoreDNS release (since 1.1.4)
 var Versions = map[string]release{
 	"1.6.6": {
 		priorVersion:   "1.6.5",
@@ -119,20 +50,20 @@ var Versions = map[string]release{
     loadbalance
 }`,
 		plugins: map[string]plugin{
-			"errors":   plugins["errors"]["v2"],
-			"log":      plugins["log"]["v1"],
-			"health":   plugins["health"]["v1"],
-			"ready":    {},
-			"autopath": {},
-			"kubernetes": plugins["kubernetes"]["v7"],
+			"errors":       plugins["errors"]["v2"],
+			"log":          plugins["log"]["v1"],
+			"health":       plugins["health"]["v1"],
+			"ready":        {},
+			"autopath":     {},
+			"kubernetes":   plugins["kubernetes"]["v7"],
 			"k8s_external": plugins["k8s_external"]["v1"],
-			"prometheus": {},
-			"forward": plugins["forward"]["v2"],
-			"cache": plugins["cache"]["v2"],
-			"loop":        {},
-			"reload":      {},
-			"loadbalance": {},
-			"hosts":       plugins["hosts"]["v1"],
+			"prometheus":   {},
+			"forward":      plugins["forward"]["v2"],
+			"cache":        plugins["cache"]["v2"],
+			"loop":         {},
+			"reload":       {},
+			"loadbalance":  {},
+			"hosts":        plugins["hosts"]["v1"],
 		},
 	},
 	"1.6.5": {
@@ -158,20 +89,20 @@ var Versions = map[string]release{
     loadbalance
 }`,
 		plugins: map[string]plugin{
-			"errors": plugins["errors"]["v2"],
-			"log":    plugins["log"]["v1"],
-			"health": plugins["health"]["v1 add lameduck"],
-			"ready":    {},
-			"autopath": {},
-			"kubernetes": plugins["kubernetes"]["v7"],
+			"errors":       plugins["errors"]["v2"],
+			"log":          plugins["log"]["v1"],
+			"health":       plugins["health"]["v1 add lameduck"],
+			"ready":        {},
+			"autopath":     {},
+			"kubernetes":   plugins["kubernetes"]["v7"],
 			"k8s_external": plugins["k8s_external"]["v1"],
-			"prometheus": {},
-			"forward": plugins["forward"]["v2"],
-			"cache": plugins["cache"]["v1"],
-			"loop":        {},
-			"reload":      {},
-			"loadbalance": {},
-			"hosts":       plugins["hosts"]["v1"],
+			"prometheus":   {},
+			"forward":      plugins["forward"]["v2"],
+			"cache":        plugins["cache"]["v1"],
+			"loop":         {},
+			"reload":       {},
+			"loadbalance":  {},
+			"hosts":        plugins["hosts"]["v1"],
 		},
 	},
 	"1.6.4": {
@@ -179,20 +110,20 @@ var Versions = map[string]release{
 		priorVersion:   "1.6.3",
 		dockerImageSHA: "493ee88e1a92abebac67cbd4b5658b4730e0f33512461442d8d9214ea6734a9b",
 		plugins: map[string]plugin{
-			"errors":   plugins["errors"]["v2"],
-			"log":      plugins["log"]["v1"],
-			"health":   plugins["health"]["v1"],
-			"ready":    {},
-			"autopath": {},
-			"kubernetes": plugins["kubernetes"]["v7"],
+			"errors":       plugins["errors"]["v2"],
+			"log":          plugins["log"]["v1"],
+			"health":       plugins["health"]["v1"],
+			"ready":        {},
+			"autopath":     {},
+			"kubernetes":   plugins["kubernetes"]["v7"],
 			"k8s_external": plugins["k8s_external"]["v1"],
-			"prometheus": {},
-			"forward": plugins["forward"]["v2"],
-			"cache": plugins["cache"]["v1"],
-			"loop":        {},
-			"reload":      {},
-			"loadbalance": {},
-			"hosts":       plugins["hosts"]["v1"],
+			"prometheus":   {},
+			"forward":      plugins["forward"]["v2"],
+			"cache":        plugins["cache"]["v1"],
+			"loop":         {},
+			"reload":       {},
+			"loadbalance":  {},
+			"hosts":        plugins["hosts"]["v1"],
 		},
 	},
 	"1.6.3": {
@@ -200,20 +131,20 @@ var Versions = map[string]release{
 		priorVersion:   "1.6.2",
 		dockerImageSHA: "cfa7236dab4e3860881fdf755880ff8361e42f6cba2e3775ae48e2d46d22f7ba",
 		plugins: map[string]plugin{
-			"errors":   plugins["errors"]["v2"],
-			"log":      plugins["log"]["v1"],
-			"health":   plugins["health"]["v1"],
-			"ready":    {},
-			"autopath": {},
-			"kubernetes": plugins["kubernetes"]["v7"],
+			"errors":       plugins["errors"]["v2"],
+			"log":          plugins["log"]["v1"],
+			"health":       plugins["health"]["v1"],
+			"ready":        {},
+			"autopath":     {},
+			"kubernetes":   plugins["kubernetes"]["v7"],
 			"k8s_external": plugins["k8s_external"]["v1"],
-			"prometheus": {},
-			"forward": plugins["forward"]["v2"],
-			"cache": plugins["cache"]["v1"],
-			"loop":        {},
-			"reload":      {},
-			"loadbalance": {},
-			"hosts":       plugins["hosts"]["v1"],
+			"prometheus":   {},
+			"forward":      plugins["forward"]["v2"],
+			"cache":        plugins["cache"]["v1"],
+			"loop":         {},
+			"reload":       {},
+			"loadbalance":  {},
+			"hosts":        plugins["hosts"]["v1"],
 		},
 	},
 	"1.6.2": {
@@ -237,20 +168,20 @@ var Versions = map[string]release{
     loadbalance
 }`,
 		plugins: map[string]plugin{
-			"errors":   plugins["errors"]["v2"],
-			"log":      plugins["log"]["v1"],
-			"health":   plugins["health"]["v1"],
-			"ready":    {},
-			"autopath": {},
-			"kubernetes": plugins["kubernetes"]["v7"],
+			"errors":       plugins["errors"]["v2"],
+			"log":          plugins["log"]["v1"],
+			"health":       plugins["health"]["v1"],
+			"ready":        {},
+			"autopath":     {},
+			"kubernetes":   plugins["kubernetes"]["v7"],
 			"k8s_external": plugins["k8s_external"]["v1"],
-			"prometheus": {},
-			"forward": plugins["forward"]["v2"],
-			"cache": plugins["cache"]["v1"],
-			"loop":        {},
-			"reload":      {},
-			"loadbalance": {},
-			"hosts":       plugins["hosts"]["v1"],
+			"prometheus":   {},
+			"forward":      plugins["forward"]["v2"],
+			"cache":        plugins["cache"]["v1"],
+			"loop":         {},
+			"reload":       {},
+			"loadbalance":  {},
+			"hosts":        plugins["hosts"]["v1"],
 		},
 	},
 	"1.6.1": {
@@ -258,20 +189,20 @@ var Versions = map[string]release{
 		priorVersion:   "1.6.0",
 		dockerImageSHA: "9ae3b6fcac4ee821362277de6bd8fd2236fa7d3e19af2ef0406d80b595620a7a",
 		plugins: map[string]plugin{
-			"errors":   plugins["errors"]["v2"],
-			"log":      plugins["log"]["v1"],
-			"health":   plugins["health"]["v1"],
-			"ready":    {},
-			"autopath": {},
-			"kubernetes": plugins["kubernetes"]["v7"],
+			"errors":       plugins["errors"]["v2"],
+			"log":          plugins["log"]["v1"],
+			"health":       plugins["health"]["v1"],
+			"ready":        {},
+			"autopath":     {},
+			"kubernetes":   plugins["kubernetes"]["v7"],
 			"k8s_external": plugins["k8s_external"]["v1"],
-			"prometheus": {},
-			"forward": plugins["forward"]["v2"],
-			"cache": plugins["cache"]["v1"],
-			"loop":        {},
-			"reload":      {},
-			"loadbalance": {},
-			"hosts":       plugins["hosts"]["v1"],
+			"prometheus":   {},
+			"forward":      plugins["forward"]["v2"],
+			"cache":        plugins["cache"]["v1"],
+			"loop":         {},
+			"reload":       {},
+			"loadbalance":  {},
+			"hosts":        plugins["hosts"]["v1"],
 		},
 	},
 	"1.6.0": {
@@ -279,20 +210,20 @@ var Versions = map[string]release{
 		priorVersion:   "1.5.2",
 		dockerImageSHA: "263d03f2b889a75a0b91e035c2a14d45d7c1559c53444c5f7abf3a76014b779d",
 		plugins: map[string]plugin{
-			"errors":   plugins["errors"]["v2"],
-			"log":      plugins["log"]["v1"],
-			"health":   plugins["health"]["v1"],
-			"ready":    {},
-			"autopath": {},
-			"kubernetes": plugins["kubernetes"]["v6"],
+			"errors":       plugins["errors"]["v2"],
+			"log":          plugins["log"]["v1"],
+			"health":       plugins["health"]["v1"],
+			"ready":        {},
+			"autopath":     {},
+			"kubernetes":   plugins["kubernetes"]["v6"],
 			"k8s_external": plugins["k8s_external"]["v1"],
-			"prometheus": {},
-			"forward": plugins["forward"]["v2"],
-			"cache": plugins["cache"]["v1"],
-			"loop":        {},
-			"reload":      {},
-			"loadbalance": {},
-			"hosts":       plugins["hosts"]["v1"],
+			"prometheus":   {},
+			"forward":      plugins["forward"]["v2"],
+			"cache":        plugins["cache"]["v1"],
+			"loop":         {},
+			"reload":       {},
+			"loadbalance":  {},
+			"hosts":        plugins["hosts"]["v1"],
 		},
 	},
 	"1.5.2": {
@@ -300,20 +231,20 @@ var Versions = map[string]release{
 		priorVersion:   "1.5.1",
 		dockerImageSHA: "586d15ec14911ee680ac9c5af20ff24b9d1412fbbf0e05862ee1f5c37baa65b2",
 		plugins: map[string]plugin{
-			"errors":   plugins["errors"]["v2"],
-			"log":      plugins["log"]["v1"],
-			"health":   plugins["health"]["v1"],
-			"ready":    {},
-			"autopath": {},
-			"kubernetes": plugins["kubernetes"]["v5"],
+			"errors":       plugins["errors"]["v2"],
+			"log":          plugins["log"]["v1"],
+			"health":       plugins["health"]["v1"],
+			"ready":        {},
+			"autopath":     {},
+			"kubernetes":   plugins["kubernetes"]["v5"],
 			"k8s_external": plugins["k8s_external"]["v1"],
-			"prometheus": {},
-			"forward": plugins["forward"]["v2"],
-			"cache": plugins["cache"]["v1"],
-			"loop":        {},
-			"reload":      {},
-			"loadbalance": {},
-			"hosts":       plugins["hosts"]["v1"],
+			"prometheus":   {},
+			"forward":      plugins["forward"]["v2"],
+			"cache":        plugins["cache"]["v1"],
+			"loop":         {},
+			"reload":       {},
+			"loadbalance":  {},
+			"hosts":        plugins["hosts"]["v1"],
 		},
 	},
 	"1.5.1": {
@@ -321,20 +252,20 @@ var Versions = map[string]release{
 		priorVersion:   "1.5.0",
 		dockerImageSHA: "451817637035535ae1fc8639753b453fa4b781d0dea557d5da5cb3c131e62ef5",
 		plugins: map[string]plugin{
-			"errors":   plugins["errors"]["v2"],
-			"log":      plugins["log"]["v1"],
-			"health":   plugins["health"]["v1"],
-			"ready":    {},
-			"autopath": {},
-			"kubernetes": plugins["kubernetes"]["v5"],
+			"errors":       plugins["errors"]["v2"],
+			"log":          plugins["log"]["v1"],
+			"health":       plugins["health"]["v1"],
+			"ready":        {},
+			"autopath":     {},
+			"kubernetes":   plugins["kubernetes"]["v5"],
 			"k8s_external": plugins["k8s_external"]["v1"],
-			"prometheus": {},
-			"forward": plugins["forward"]["v2"],
-			"cache": plugins["cache"]["v1"],
-			"loop":        {},
-			"reload":      {},
-			"loadbalance": {},
-			"hosts":       plugins["hosts"]["v1"],
+			"prometheus":   {},
+			"forward":      plugins["forward"]["v2"],
+			"cache":        plugins["cache"]["v1"],
+			"loop":         {},
+			"reload":       {},
+			"loadbalance":  {},
+			"hosts":        plugins["hosts"]["v1"],
 		},
 	},
 	"1.5.0": {
@@ -344,7 +275,7 @@ var Versions = map[string]release{
 		plugins: map[string]plugin{
 			"errors": plugins["errors"]["v2"],
 			"log":    plugins["log"]["v1"],
-			"health":   plugins["health"]["v1"],
+			"health": plugins["health"]["v1"],
 			"ready": {
 				status: newdefault,
 				add: func(c *corefile.Server) (*corefile.Server, error) {
@@ -352,22 +283,17 @@ var Versions = map[string]release{
 				},
 				downAction: removePlugin,
 			},
-			"autopath": {},
-			"kubernetes": plugins["kubernetes"]["v5"],
+			"autopath":     {},
+			"kubernetes":   plugins["kubernetes"]["v5"],
 			"k8s_external": plugins["k8s_external"]["v1"],
-			"prometheus": {},
-			"proxy": {
-				status:       removed,
-				replacedBy:   "forward",
-				action:       proxyToForwardPluginAction,
-				namedOptions: proxyToForwardOptionsMigrations,
-			},
-			"forward": plugins["forward"]["v2"],
-			"cache": plugins["cache"]["v1"],
-			"loop":        {},
-			"reload":      {},
-			"loadbalance": {},
-			"hosts":       plugins["hosts"]["v1"],
+			"prometheus":   {},
+			"proxy":        plugins["proxy"]["removal"],
+			"forward":      plugins["forward"]["v2"],
+			"cache":        plugins["cache"]["v1"],
+			"loop":         {},
+			"reload":       {},
+			"loadbalance":  {},
+			"hosts":        plugins["hosts"]["v1"],
 		},
 		postProcess: breakForwardStubDomainsIntoServerBlocks,
 	},
@@ -376,25 +302,20 @@ var Versions = map[string]release{
 		priorVersion:   "1.3.1",
 		dockerImageSHA: "70a92e9f6fc604f9b629ca331b6135287244a86612f550941193ec7e12759417",
 		plugins: map[string]plugin{
-			"errors":   plugins["errors"]["v2"],
-			"log":      plugins["log"]["v1"],
-			"health":   plugins["health"]["v1"],
-			"autopath": {},
-			"kubernetes": plugins["kubernetes"]["v4"],
+			"errors":       plugins["errors"]["v2"],
+			"log":          plugins["log"]["v1"],
+			"health":       plugins["health"]["v1"],
+			"autopath":     {},
+			"kubernetes":   plugins["kubernetes"]["v4"],
 			"k8s_external": plugins["k8s_external"]["v1"],
-			"prometheus": {},
-			"proxy": {
-				status:       deprecated,
-				replacedBy:   "forward",
-				action:       proxyToForwardPluginAction,
-				namedOptions: proxyToForwardOptionsMigrations,
-			},
-			"forward": plugins["forward"]["v2"],
-			"cache": plugins["cache"]["v1"],
-			"loop":        {},
-			"reload":      {},
-			"loadbalance": {},
-			"hosts":       plugins["hosts"]["v1"],
+			"prometheus":   {},
+			"proxy":        plugins["proxy"]["deprecation"],
+			"forward":      plugins["forward"]["v2"],
+			"cache":        plugins["cache"]["v1"],
+			"loop":         {},
+			"reload":       {},
+			"loadbalance":  {},
+			"hosts":        plugins["hosts"]["v1"],
 		},
 		postProcess: breakForwardStubDomainsIntoServerBlocks,
 	},
@@ -420,30 +341,20 @@ var Versions = map[string]release{
     loadbalance
 }`,
 		plugins: map[string]plugin{
-			"errors":   plugins["errors"]["v2"],
-			"log":      plugins["log"]["v1"],
-			"health":   plugins["health"]["v1"],
-			"autopath": {},
-			"kubernetes": plugins["kubernetes"]["v3"],
+			"errors":       plugins["errors"]["v2"],
+			"log":          plugins["log"]["v1"],
+			"health":       plugins["health"]["v1"],
+			"autopath":     {},
+			"kubernetes":   plugins["kubernetes"]["v3"],
 			"k8s_external": plugins["k8s_external"]["v1"],
-			"prometheus": {},
-			"proxy": {
-				namedOptions: map[string]option{
-					"policy":       {},
-					"fail_timeout": {},
-					"max_fails":    {},
-					"health_check": {},
-					"except":       {},
-					"spray":        {},
-					"protocol":     {},
-				},
-			},
-			"forward": plugins["forward"]["v2"],
-			"cache": plugins["cache"]["v1"],
-			"loop":        {},
-			"reload":      {},
-			"loadbalance": {},
-			"hosts":       plugins["hosts"]["v1"],
+			"prometheus":   {},
+			"proxy":        plugins["proxy"]["v2"],
+			"forward":      plugins["forward"]["v2"],
+			"cache":        plugins["cache"]["v1"],
+			"loop":         {},
+			"reload":       {},
+			"loadbalance":  {},
+			"hosts":        plugins["hosts"]["v1"],
 		},
 	},
 	"1.3.0": {
@@ -451,30 +362,20 @@ var Versions = map[string]release{
 		priorVersion:   "1.2.6",
 		dockerImageSHA: "e030773c7fee285435ed7fc7623532ee54c4c1c4911fb24d95cd0170a8a768bc",
 		plugins: map[string]plugin{
-			"errors":   plugins["errors"]["v2"],
-			"log":      plugins["log"]["v1"],
-			"health":   plugins["health"]["v1"],
-			"autopath": {},
-			"kubernetes": plugins["kubernetes"]["v2"],
+			"errors":       plugins["errors"]["v2"],
+			"log":          plugins["log"]["v1"],
+			"health":       plugins["health"]["v1"],
+			"autopath":     {},
+			"kubernetes":   plugins["kubernetes"]["v2"],
 			"k8s_external": plugins["k8s_external"]["v1"],
-			"prometheus": {},
-			"proxy": {
-				namedOptions: map[string]option{
-					"policy":       {},
-					"fail_timeout": {},
-					"max_fails":    {},
-					"health_check": {},
-					"except":       {},
-					"spray":        {},
-					"protocol":     {},
-				},
-			},
-			"forward": plugins["forward"]["v2"],
-			"cache": plugins["cache"]["v1"],
-			"loop":        {},
-			"reload":      {},
-			"loadbalance": {},
-			"hosts":       plugins["hosts"]["v1"],
+			"prometheus":   {},
+			"proxy":        plugins["proxy"]["v2"],
+			"forward":      plugins["forward"]["v2"],
+			"cache":        plugins["cache"]["v1"],
+			"loop":         {},
+			"reload":       {},
+			"loadbalance":  {},
+			"hosts":        plugins["hosts"]["v1"],
 		},
 	},
 	"1.2.6": {
@@ -503,24 +404,14 @@ var Versions = map[string]release{
 					"consolidate": {},
 				},
 			},
-			"log":      plugins["log"]["v1"],
-			"health":   plugins["health"]["v1"],
-			"autopath": {},
-			"kubernetes": plugins["kubernetes"]["v2"],
-			"prometheus": {},
-			"proxy": {
-				namedOptions: map[string]option{
-					"policy":       {},
-					"fail_timeout": {},
-					"max_fails":    {},
-					"health_check": {},
-					"except":       {},
-					"spray":        {},
-					"protocol":     {},
-				},
-			},
-			"forward": plugins["forward"]["v2"],
-			"cache": plugins["cache"]["v1"],
+			"log":         plugins["log"]["v1"],
+			"health":      plugins["health"]["v1"],
+			"autopath":    {},
+			"kubernetes":  plugins["kubernetes"]["v2"],
+			"prometheus":  {},
+			"proxy":       plugins["proxy"]["v2"],
+			"forward":     plugins["forward"]["v2"],
+			"cache":       plugins["cache"]["v1"],
 			"loop":        {},
 			"reload":      {},
 			"loadbalance": {},
@@ -532,25 +423,15 @@ var Versions = map[string]release{
 		priorVersion:   "1.2.4",
 		dockerImageSHA: "33c8da20b887ae12433ec5c40bfddefbbfa233d5ce11fb067122e68af30291d6",
 		plugins: map[string]plugin{
-			"errors":   plugins["errors"]["v1"],
-			"log":      plugins["log"]["v1"],
-			"health":   plugins["health"]["v1"],
-			"autopath": {},
-			"kubernetes": plugins["kubernetes"]["v2"],
-			"prometheus": {},
-			"proxy": {
-				namedOptions: map[string]option{
-					"policy":       {},
-					"fail_timeout": {},
-					"max_fails":    {},
-					"health_check": {},
-					"except":       {},
-					"spray":        {},
-					"protocol":     {},
-				},
-			},
-			"forward": plugins["forward"]["v2"],
-			"cache": plugins["cache"]["v1"],
+			"errors":      plugins["errors"]["v1"],
+			"log":         plugins["log"]["v1"],
+			"health":      plugins["health"]["v1"],
+			"autopath":    {},
+			"kubernetes":  plugins["kubernetes"]["v2"],
+			"prometheus":  {},
+			"proxy":       plugins["proxy"]["v2"],
+			"forward":     plugins["forward"]["v2"],
+			"cache":       plugins["cache"]["v1"],
 			"loop":        {},
 			"reload":      {},
 			"loadbalance": {},
@@ -562,25 +443,15 @@ var Versions = map[string]release{
 		priorVersion:   "1.2.3",
 		dockerImageSHA: "a0d40ad961a714c699ee7b61b77441d165f6252f9fb84ac625d04a8d8554c0ec",
 		plugins: map[string]plugin{
-			"errors":   plugins["errors"]["v1"],
-			"log":      plugins["log"]["v1"],
-			"health":   plugins["health"]["v1"],
-			"autopath": {},
-			"kubernetes": plugins["kubernetes"]["v2"],
-			"prometheus": {},
-			"proxy": {
-				namedOptions: map[string]option{
-					"policy":       {},
-					"fail_timeout": {},
-					"max_fails":    {},
-					"health_check": {},
-					"except":       {},
-					"spray":        {},
-					"protocol":     {},
-				},
-			},
-			"forward": plugins["forward"]["v2"],
-			"cache": plugins["cache"]["v1"],
+			"errors":      plugins["errors"]["v1"],
+			"log":         plugins["log"]["v1"],
+			"health":      plugins["health"]["v1"],
+			"autopath":    {},
+			"kubernetes":  plugins["kubernetes"]["v2"],
+			"prometheus":  {},
+			"proxy":       plugins["proxy"]["v2"],
+			"forward":     plugins["forward"]["v2"],
+			"cache":       plugins["cache"]["v1"],
 			"loop":        {},
 			"reload":      {},
 			"loadbalance": {},
@@ -592,25 +463,15 @@ var Versions = map[string]release{
 		priorVersion:   "1.2.2",
 		dockerImageSHA: "12f3cab301c826978fac736fd40aca21ac023102fd7f4aa6b4341ae9ba89e90e",
 		plugins: map[string]plugin{
-			"errors":   plugins["errors"]["v1"],
-			"log":      plugins["log"]["v1"],
-			"health":   plugins["health"]["v1"],
-			"autopath": {},
-			"kubernetes": plugins["kubernetes"]["v2"],
-			"prometheus": {},
-			"proxy": {
-				namedOptions: map[string]option{
-					"policy":       {},
-					"fail_timeout": {},
-					"max_fails":    {},
-					"health_check": {},
-					"except":       {},
-					"spray":        {},
-					"protocol":     {},
-				},
-			},
-			"forward": plugins["forward"]["v2"],
-			"cache": plugins["cache"]["v1"],
+			"errors":      plugins["errors"]["v1"],
+			"log":         plugins["log"]["v1"],
+			"health":      plugins["health"]["v1"],
+			"autopath":    {},
+			"kubernetes":  plugins["kubernetes"]["v2"],
+			"prometheus":  {},
+			"proxy":       plugins["proxy"]["v2"],
+			"forward":     plugins["forward"]["v2"],
+			"cache":       plugins["cache"]["v1"],
 			"loop":        {},
 			"reload":      {},
 			"loadbalance": {},
@@ -638,25 +499,15 @@ var Versions = map[string]release{
     loadbalance
 }`,
 		plugins: map[string]plugin{
-			"errors":   plugins["errors"]["v1"],
-			"log":      plugins["log"]["v1"],
-			"health":   plugins["health"]["v1"],
-			"autopath": {},
-			"kubernetes": plugins["kubernetes"]["v1"],
-			"prometheus": {},
-			"proxy": {
-				namedOptions: map[string]option{
-					"policy":       {},
-					"fail_timeout": {},
-					"max_fails":    {},
-					"health_check": {},
-					"except":       {},
-					"spray":        {},
-					"protocol":     {},
-				},
-			},
-			"forward": plugins["forward"]["v2"],
-			"cache": plugins["cache"]["v1"],
+			"errors":      plugins["errors"]["v1"],
+			"log":         plugins["log"]["v1"],
+			"health":      plugins["health"]["v1"],
+			"autopath":    {},
+			"kubernetes":  plugins["kubernetes"]["v1"],
+			"prometheus":  {},
+			"proxy":       plugins["proxy"]["v2"],
+			"forward":     plugins["forward"]["v2"],
+			"cache":       plugins["cache"]["v1"],
 			"loop":        {},
 			"reload":      {},
 			"loadbalance": {},
@@ -668,25 +519,15 @@ var Versions = map[string]release{
 		priorVersion:   "1.2.0",
 		dockerImageSHA: "fb129c6a7c8912bc6d9cc4505e1f9007c5565ceb1aa6369750e60cc79771a244",
 		plugins: map[string]plugin{
-			"errors":   plugins["errors"]["v1"],
-			"log":      plugins["log"]["v1"],
-			"health":   plugins["health"]["v1"],
-			"autopath": {},
+			"errors":     plugins["errors"]["v1"],
+			"log":        plugins["log"]["v1"],
+			"health":     plugins["health"]["v1"],
+			"autopath":   {},
 			"kubernetes": plugins["kubernetes"]["v1"],
 			"prometheus": {},
-			"proxy": {
-				namedOptions: map[string]option{
-					"policy":       {},
-					"fail_timeout": {},
-					"max_fails":    {},
-					"health_check": {},
-					"except":       {},
-					"spray":        {},
-					"protocol":     {},
-				},
-			},
-			"forward": plugins["forward"]["v2"],
-			"cache": plugins["cache"]["v1"],
+			"proxy":      plugins["proxy"]["v2"],
+			"forward":    plugins["forward"]["v2"],
+			"cache":      plugins["cache"]["v1"],
 			"loop": {
 				status: newdefault,
 				add: func(s *corefile.Server) (*corefile.Server, error) {
@@ -704,28 +545,15 @@ var Versions = map[string]release{
 		priorVersion:   "1.1.4",
 		dockerImageSHA: "ae69a32f8cc29a3e2af9628b6473f24d3e977950a2cb62ce8911478a61215471",
 		plugins: map[string]plugin{
-			"errors":   plugins["errors"]["v1"],
-			"log":      plugins["log"]["v1"],
-			"health":   plugins["health"]["v1"],
-			"autopath": {},
-			"kubernetes": plugins["kubernetes"]["v1"],
-			"prometheus": {},
-			"proxy": {
-				namedOptions: map[string]option{
-					"policy":       {},
-					"fail_timeout": {},
-					"max_fails":    {},
-					"health_check": {},
-					"except":       {},
-					"spray":        {},
-					"protocol": {
-						status: removed,
-						action: proxyRemoveHttpsGoogleProtocol,
-					},
-				},
-			},
-			"forward": plugins["forward"]["v2"],
-			"cache": plugins["cache"]["v1"],
+			"errors":      plugins["errors"]["v1"],
+			"log":         plugins["log"]["v1"],
+			"health":      plugins["health"]["v1"],
+			"autopath":    {},
+			"kubernetes":  plugins["kubernetes"]["v1"],
+			"prometheus":  {},
+			"proxy":       plugins["proxy"]["v2"],
+			"forward":     plugins["forward"]["v2"],
+			"cache":       plugins["cache"]["v1"],
 			"reload":      {},
 			"loadbalance": {},
 			"hosts":       plugins["hosts"]["v1"],
@@ -736,28 +564,15 @@ var Versions = map[string]release{
 		priorVersion:   "1.1.3",
 		dockerImageSHA: "463c7021141dd3bfd4a75812f4b735ef6aadc0253a128f15ffe16422abe56e50",
 		plugins: map[string]plugin{
-			"errors":   plugins["errors"]["v1"],
-			"log":      plugins["log"]["v1"],
-			"health":   plugins["health"]["v1"],
-			"autopath": {},
-			"kubernetes": plugins["kubernetes"]["v1"],
-			"prometheus": {},
-			"proxy": {
-				namedOptions: map[string]option{
-					"policy":       {},
-					"fail_timeout": {},
-					"max_fails":    {},
-					"health_check": {},
-					"except":       {},
-					"spray":        {},
-					"protocol": {
-						status: ignored,
-						action: proxyRemoveHttpsGoogleProtocol,
-					},
-				},
-			},
-			"forward": plugins["forward"]["v1"],
-			"cache": plugins["cache"]["v1"],
+			"errors":      plugins["errors"]["v1"],
+			"log":         plugins["log"]["v1"],
+			"health":      plugins["health"]["v1"],
+			"autopath":    {},
+			"kubernetes":  plugins["kubernetes"]["v1"],
+			"prometheus":  {},
+			"proxy":       plugins["proxy"]["v1"],
+			"forward":     plugins["forward"]["v1"],
+			"cache":       plugins["cache"]["v1"],
 			"reload":      {},
 			"loadbalance": {},
 			"hosts":       plugins["hosts"]["v1"],
@@ -780,90 +595,4 @@ var Versions = map[string]release{
     cache 30
     reload
 }`},
-}
-
-var proxyToForwardOptionsMigrations = map[string]option{
-	"policy": {
-		action: func(o *corefile.Option) (*corefile.Option, error) {
-			if len(o.Args) == 2 && o.Args[1] == "least_conn" {
-				o.Name = "force_tcp"
-				o.Args = nil
-			}
-			return o, nil
-		},
-	},
-	"except":       {},
-	"fail_timeout": {action: removeOption},
-	"max_fails":    {action: removeOption},
-	"health_check": {action: removeOption},
-	"spray":        {action: removeOption},
-	"protocol": {
-		action: func(o *corefile.Option) (*corefile.Option, error) {
-			if len(o.Args) >= 2 && o.Args[1] == "force_tcp" {
-				o.Name = "force_tcp"
-				o.Args = nil
-				return o, nil
-			}
-			return nil, nil
-		},
-	},
-}
-
-var proxyToForwardPluginAction = func(p *corefile.Plugin) (*corefile.Plugin, error) {
-	return renamePlugin(p, "forward")
-}
-
-var useFirstArgumentOnly = func(o *corefile.Option) (*corefile.Option, error) {
-	if len(o.Args) < 1 {
-		return o, nil
-	}
-	o.Args = o.Args[:1]
-	return o, nil
-}
-
-var proxyRemoveHttpsGoogleProtocol = func(o *corefile.Option) (*corefile.Option, error) {
-	if len(o.Args) > 0 && o.Args[0] == "https_google" {
-		return nil, nil
-	}
-	return o, nil
-}
-
-func breakForwardStubDomainsIntoServerBlocks(cf *corefile.Corefile) (*corefile.Corefile, error) {
-	for _, sb := range cf.Servers {
-		for j, fwd := range sb.Plugins {
-			if fwd.Name != "forward" {
-				continue
-			}
-			if len(fwd.Args) == 0 {
-				return nil, errors.New("found invalid forward plugin declaration")
-			}
-			if fwd.Args[0] == "." {
-				// dont move the default upstream
-				continue
-			}
-			if len(sb.DomPorts) != 1 {
-				return cf, errors.New("unhandled migration of multi-domain/port server block")
-			}
-			if sb.DomPorts[0] != "." && sb.DomPorts[0] != ".:53" {
-				return cf, errors.New("unhandled migration of non-default domain/port server block")
-			}
-
-			newSb := &corefile.Server{}                // create a new server block
-			newSb.DomPorts = []string{fwd.Args[0]}     // copy the forward zone to the server block domain
-			fwd.Args[0] = "."                          // the plugin's zone changes to "." for brevity
-			newSb.Plugins = append(newSb.Plugins, fwd) // add the plugin to its new server block
-
-			// Add appropriate addtl plugins to new server block
-			newSb.Plugins = append(newSb.Plugins, &corefile.Plugin{Name: "loop"})
-			newSb.Plugins = append(newSb.Plugins, &corefile.Plugin{Name: "errors"})
-			newSb.Plugins = append(newSb.Plugins, &corefile.Plugin{Name: "cache", Args: []string{"30"}})
-
-			//add new server block to corefile
-			cf.Servers = append(cf.Servers, newSb)
-
-			//remove the forward plugin from the original server block
-			sb.Plugins = append(sb.Plugins[:j], sb.Plugins[j+1:]...)
-		}
-	}
-	return cf, nil
 }
